requirements.txt
python-telegram-bot[job-queue]>=22.3
SQLAlchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
aiohttp>=3.9.5
pydantic>=2.7.0
python-dotenv>=1.0.1
